move etools-toast
