+++ src_ts/components/app-shell/app-shell.ts
@@ -57,7 +57,7 @@ import LoadingMixin from '@unicef-polymer/etools-loading/etools-loading-mixin.js
 import 'etools-piwik-analytics/etools-piwik-analytics.js';
 import {AppMenuMixin} from './menu/mixins/app-menu-mixin.js';
 import CommonDataMixin from '../common/common-data.js';
-import ToastNotificationsMixin from '../toast-notifications/toast-notification-mixin.js';
+import ToastNotificationsMixin from '../common/toast-notifications/toast-notification-mixin.js';
 import ScrollControlMixin from '../common/mixins/scroll-control-mixin.js';
 import UserDataMixin from '../user/user-data-mixin';
